requirements.txt
camoufox
python-dotenv
